working v1 of decoupled
